CoreX/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
dj-database-url==2.3.0
Django==5.1.5
django-nine==0.2.7
djangorestframework==3.15.2
gunicorn==23.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==1.26.20
whitenoise==6.8.2
